+++ coderd/organizations_test.go
@@ -7,7 +7,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/coder/coder/v2/coderd/coderdtest"
-	"github.com/coder/coder/v2/coderd/util/ptr"
 	"github.com/coder/coder/v2/codersdk"
 	"github.com/coder/coder/v2/testutil"
 )
@@ -122,223 +121,5 @@ func TestPostOrganizationsByUser(t *testing.T) {
 		require.NoError(t, err)
 		require.Equal(t, "new-org", o.Name)
 		require.Equal(t, "new-org", o.DisplayName) // should match the given `Name`
-	})
-}
-
-func TestPatchOrganizationsByUser(t *testing.T) {
-	t.Parallel()
-	t.Run("Conflict", func(t *testing.T) {
-		t.Parallel()
-		client := coderdtest.New(t, nil)
-		user := coderdtest.CreateFirstUser(t, client)
-		ctx := testutil.Context(t, testutil.WaitMedium)
-
-		originalOrg, err := client.Organization(ctx, user.OrganizationID)
-		require.NoError(t, err)
-		o, err := client.CreateOrganization(ctx, codersdk.CreateOrganizationRequest{
-			Name:        "something-unique",
-			DisplayName: "Something Unique",
-		})
-		require.NoError(t, err)
-
-		_, err = client.UpdateOrganization(ctx, o.ID.String(), codersdk.UpdateOrganizationRequest{
-			Name: originalOrg.Name,
-		})
-		var apiErr *codersdk.Error
-		require.ErrorAs(t, err, &apiErr)
-		require.Equal(t, http.StatusConflict, apiErr.StatusCode())
-	})
-
-	t.Run("ReservedName", func(t *testing.T) {
-		t.Parallel()
-		client := coderdtest.New(t, nil)
-		_ = coderdtest.CreateFirstUser(t, client)
-		ctx := testutil.Context(t, testutil.WaitMedium)
-
-		o, err := client.CreateOrganization(ctx, codersdk.CreateOrganizationRequest{
-			Name:        "something-unique",
-			DisplayName: "Something Unique",
-		})
-		require.NoError(t, err)
-
-		_, err = client.UpdateOrganization(ctx, o.ID.String(), codersdk.UpdateOrganizationRequest{
-			Name: codersdk.DefaultOrganization,
-		})
-		var apiErr *codersdk.Error
-		require.ErrorAs(t, err, &apiErr)
-		require.Equal(t, http.StatusBadRequest, apiErr.StatusCode())
-	})
-
-	t.Run("InvalidName", func(t *testing.T) {
-		t.Parallel()
-		client := coderdtest.New(t, nil)
-		_ = coderdtest.CreateFirstUser(t, client)
-		ctx := testutil.Context(t, testutil.WaitMedium)
-
-		o, err := client.CreateOrganization(ctx, codersdk.CreateOrganizationRequest{
-			Name:        "something-unique",
-			DisplayName: "Something Unique",
-		})
-		require.NoError(t, err)
-
-		_, err = client.UpdateOrganization(ctx, o.ID.String(), codersdk.UpdateOrganizationRequest{
-			Name: "something unique but not url safe",
-		})
-		var apiErr *codersdk.Error
-		require.ErrorAs(t, err, &apiErr)
-		require.Equal(t, http.StatusBadRequest, apiErr.StatusCode())
-	})
-
-	t.Run("UpdateById", func(t *testing.T) {
-		t.Parallel()
-		client := coderdtest.New(t, nil)
-		_ = coderdtest.CreateFirstUser(t, client)
-		ctx := testutil.Context(t, testutil.WaitMedium)
-
-		o, err := client.CreateOrganization(ctx, codersdk.CreateOrganizationRequest{
-			Name:        "new-org",
-			DisplayName: "New organization",
-		})
-		require.NoError(t, err)
-
-		o, err = client.UpdateOrganization(ctx, o.ID.String(), codersdk.UpdateOrganizationRequest{
-			Name: "new-new-org",
-		})
-		require.NoError(t, err)
-		require.Equal(t, "new-new-org", o.Name)
-	})
-
-	t.Run("UpdateByName", func(t *testing.T) {
-		t.Parallel()
-		client := coderdtest.New(t, nil)
-		_ = coderdtest.CreateFirstUser(t, client)
-		ctx := testutil.Context(t, testutil.WaitMedium)
-
-		o, err := client.CreateOrganization(ctx, codersdk.CreateOrganizationRequest{
-			Name:        "new-org",
-			DisplayName: "New organization",
-		})
-		require.NoError(t, err)
-
-		o, err = client.UpdateOrganization(ctx, o.Name, codersdk.UpdateOrganizationRequest{
-			Name: "new-new-org",
-		})
-		require.NoError(t, err)
-		require.Equal(t, "new-new-org", o.Name)
-		require.Equal(t, "New organization", o.DisplayName) // didn't change
-	})
-
-	t.Run("UpdateDisplayName", func(t *testing.T) {
-		t.Parallel()
-		client := coderdtest.New(t, nil)
-		_ = coderdtest.CreateFirstUser(t, client)
-		ctx := testutil.Context(t, testutil.WaitMedium)
-
-		o, err := client.CreateOrganization(ctx, codersdk.CreateOrganizationRequest{
-			Name:        "new-org",
-			DisplayName: "New organization",
-		})
-		require.NoError(t, err)
-
-		o, err = client.UpdateOrganization(ctx, o.Name, codersdk.UpdateOrganizationRequest{
-			DisplayName: "The Newest One",
-		})
-		require.NoError(t, err)
-		require.Equal(t, "new-org", o.Name) // didn't change
-		require.Equal(t, "The Newest One", o.DisplayName)
-	})
-
-	t.Run("UpdateDescription", func(t *testing.T) {
-		t.Parallel()
-		client := coderdtest.New(t, nil)
-		_ = coderdtest.CreateFirstUser(t, client)
-		ctx := testutil.Context(t, testutil.WaitMedium)
-
-		o, err := client.CreateOrganization(ctx, codersdk.CreateOrganizationRequest{
-			Name:        "new-org",
-			DisplayName: "New organization",
-		})
-		require.NoError(t, err)
-
-		o, err = client.UpdateOrganization(ctx, o.Name, codersdk.UpdateOrganizationRequest{
-			Description: ptr.Ref("wow, this organization description is so updated!"),
-		})
-
-		require.NoError(t, err)
-		require.Equal(t, "new-org", o.Name)                 // didn't change
-		require.Equal(t, "New organization", o.DisplayName) // didn't change
-		require.Equal(t, "wow, this organization description is so updated!", o.Description)
-	})
-
-	t.Run("UpdateIcon", func(t *testing.T) {
-		t.Parallel()
-		client := coderdtest.New(t, nil)
-		_ = coderdtest.CreateFirstUser(t, client)
-		ctx := testutil.Context(t, testutil.WaitMedium)
-
-		o, err := client.CreateOrganization(ctx, codersdk.CreateOrganizationRequest{
-			Name:        "new-org",
-			DisplayName: "New organization",
-		})
-		require.NoError(t, err)
-
-		o, err = client.UpdateOrganization(ctx, o.Name, codersdk.UpdateOrganizationRequest{
-			Icon: ptr.Ref("/emojis/1f48f-1f3ff.png"),
-		})
-
-		require.NoError(t, err)
-		require.Equal(t, "new-org", o.Name)                 // didn't change
-		require.Equal(t, "New organization", o.DisplayName) // didn't change
-		require.Equal(t, "/emojis/1f48f-1f3ff.png", o.Icon)
-	})
-}
-
-func TestDeleteOrganizationsByUser(t *testing.T) {
-	t.Parallel()
-	t.Run("Default", func(t *testing.T) {
-		t.Parallel()
-		client := coderdtest.New(t, nil)
-		user := coderdtest.CreateFirstUser(t, client)
-		ctx := testutil.Context(t, testutil.WaitMedium)
-
-		o, err := client.Organization(ctx, user.OrganizationID)
-		require.NoError(t, err)
-
-		err = client.DeleteOrganization(ctx, o.ID.String())
-		var apiErr *codersdk.Error
-		require.ErrorAs(t, err, &apiErr)
-		require.Equal(t, http.StatusBadRequest, apiErr.StatusCode())
-	})
-
-	t.Run("DeleteById", func(t *testing.T) {
-		t.Parallel()
-		client := coderdtest.New(t, nil)
-		_ = coderdtest.CreateFirstUser(t, client)
-		ctx := testutil.Context(t, testutil.WaitMedium)
-
-		o, err := client.CreateOrganization(ctx, codersdk.CreateOrganizationRequest{
-			Name:        "doomed",
-			DisplayName: "Doomed",
-		})
-		require.NoError(t, err)
-
-		err = client.DeleteOrganization(ctx, o.ID.String())
-		require.NoError(t, err)
-	})
-
-	t.Run("DeleteByName", func(t *testing.T) {
-		t.Parallel()
-		client := coderdtest.New(t, nil)
-		_ = coderdtest.CreateFirstUser(t, client)
-		ctx := testutil.Context(t, testutil.WaitMedium)
-
-		o, err := client.CreateOrganization(ctx, codersdk.CreateOrganizationRequest{
-			Name:        "doomed",
-			DisplayName: "Doomed",
-		})
-		require.NoError(t, err)
-
-		err = client.DeleteOrganization(ctx, o.Name)
-		require.NoError(t, err)
 	})
 }
